requirements.txt
numpy==1.17.2
pandas==0.25.1
joblib==0.14.0
scikit-learn==0.21.3
scipy==1.3.1
sklearn==0.0
tornado==6.0.3
PyYAML==5.1.2
